chore: move wasm functions structure function to functiondef
